# swagger.yaml
openapi: 3.0.3
info:
  title: News API
  version: 1.0.1
  description: API for retrieving and deleting news items, with filtering capabilities.
servers:
  - url: /api/v1
    description: Base URL for v1 API

paths:
  /news:
    get:
      summary: Get all news
      operationId: getAllNews
      tags:
        - News
      parameters:
        - name: author
          in: query
          schema:
            type: string
          description: Filter by author name
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
          description: Filter by one or more tags (comma-separated)
        - name: title
          in: query
          schema:
            type: string
          description: Filter by story title
        - name: month
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: Filter by month of creation
      responses:
        '200':
          description: A list of news items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NewsItem'
        '400':
          description: Invalid filter format
        '500':
          description: Server error

  /news/{object_id}:
    get:
      summary: Get a single news item by object_id
      operationId: getNewsById
      tags:
        - News
      parameters:
        - name: object_id
          in: path
          required: false
          schema:
            type: string
      responses:
        '200':
          description: The requested news item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItem'
        '404':
          description: News item not found
        '500':
          description: Server error

    delete:
      summary: Delete a news item by object_id
      operationId: deleteNewsById
      tags:
        - News
      parameters:
        - name: object_id
          in: path
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Successfully deleted (no content)
        '404':
          description: News item not found
        '500':
          description: Server error
  /api/status:
    get:
      summary: Health check
      description: Public endpoint without required token.
      responses:
        "200":
          description: Ok add request

components:
  securitySchemes:
    ServiceTokenHeader:
      type: apiKey
      in: header
      name: X-Service-Token
      description: JWT token used for internal communication

  schemas:
    NewsItem:
      type: object
      properties:
        author:
          type: string
          example: "John Doe"
        comment_text:
          type: string
          example: "Interesting article!"
        story_title:
          type: string
          example: "Breaking News: AI Takes Over"
        story_url:
          type: string
          format: uri
          example: "https://example.com/story"
        tags:
          type: array
          items:
            type: string
          example: ["AI", "Technology", "Trends"]
        created_at:
          type: string
          format: date-time
          example: "2025-10-29T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-10-30T08:15:00Z"
        object_id:
          type: string
          example: "abc123"
        parent_id:
          type: integer
          format: int64
          example: 987654321
        story_id:
          type: integer
          format: int64
          example: 123456789
security:
  - ServiceTokenHeader: []